:sparkles: feat : 보유내역 조회 기능 추가 및 관련 DTO, Mapper, Service 수정

+++ src/main/java/com/finsight/backend/domain/vo/user/HoldingsVO.java
@@ -4,6 +4,7 @@ import com.finsight.backend.dto.request.TradeRequest;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,7 +14,7 @@ public class HoldingsVO {
     private String userId;
     private String productCode;
     private String productCategory;
-    private Long holdingsTotalPrice;
+    private BigDecimal holdingsTotalPrice;
     private Integer holdingsTotalQuantity;
     private String holdingsStatus; // holding, zero, terminated, expired
