+++ src/main/java/com/finsight/backend/tmpdetailhodings/dto/DetailHoldingsResponseDto.java
@@ -2,6 +2,7 @@ package com.finsight.backend.tmpdetailhodings.dto;
 
 import com.finsight.backend.tmpdetailhodings.vo.DetailHistoryVO;
 import lombok.*;
+import java.math.BigDecimal;
 
 import java.util.List;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -17,7 +18,7 @@ public class DetailHoldingsResponseDto {
 
     private String productCategory;
 
-    private double holdingsTotalPrice;
+    private BigDecimal holdingsTotalPrice;
 
     private int holdingsTotalQuantity;
