+++ src/main/java/com/finsight/backend/tmpdetailhodings/vo/DetailHistoryVO.java
@@ -1,7 +1,7 @@
 package com.finsight.backend.tmpdetailhodings.vo;
 
 import lombok.Data;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -16,5 +16,5 @@ public class DetailHistoryVO {
     private LocalDateTime historyTradeDate;
 
     private int historyQuantity;
-    private Long historyAmount;
+    private BigDecimal historyAmount;
 }
